sources:vector:refactor
