+++ src/sources/vector/fetch/compression.rs
