+++ src/sources/vector/fetch/service.rs
@@ -1,4 +1,3 @@
-use snafu::Snafu;
 use std::task::{Context, Poll};
 
 use super::compression::VectorCompression;
@@ -7,62 +6,15 @@ use http::Uri;
 use hyper::client::HttpConnector;
 use hyper_openssl::HttpsConnector;
 use hyper_proxy::ProxyConnector;
-use prost::Message;
-use tonic::{IntoRequest, body::BoxBody};
-use tower::Service;
-use vector_lib::request_metadata::{MetaDescriptive, RequestMetadata};
-
+use tonic::{body::BoxBody};
 use crate::{
     Error,
-    event::{EventFinalizers, Finalizable},
-    internal_events::EndpointBytesSent,
     proto::vector as proto_vector,
-    sinks::util::uri,
 };
-
-#[derive(Debug, Snafu)]
-#[snafu(visibility(pub))]
-pub enum VectorSinkError {
-    #[snafu(display("Request failed: {}", source))]
-    Request { source: tonic::Status },
-
-    #[snafu(display("Vector source unhealthy: {:?}", status))]
-    Health { status: Option<&'static str> },
-
-    #[snafu(display("URL has no host."))]
-    NoHost,
-}
 
 #[derive(Clone, Debug)]
 pub struct VectorService {
     pub client: proto_vector::Client<HyperSvc>,
-    pub protocol: String,
-    pub endpoint: String,
-}
-
-pub type VectorResponse = tonic::Streaming<crate::proto::vector::PullEventsResponse>;
-
-#[derive(Clone, Default)]
-pub struct VectorRequest {
-    pub finalizers: EventFinalizers,
-    pub metadata: RequestMetadata,
-    pub request: proto_vector::PullEventsRequest,
-}
-
-impl Finalizable for VectorRequest {
-    fn take_finalizers(&mut self) -> EventFinalizers {
-        self.finalizers.take_finalizers()
-    }
-}
-
-impl MetaDescriptive for VectorRequest {
-    fn get_metadata(&self) -> &RequestMetadata {
-        &self.metadata
-    }
-
-    fn metadata_mut(&mut self) -> &mut RequestMetadata {
-        &mut self.metadata
-    }
 }
 
 impl VectorService {
@@ -71,7 +23,6 @@ impl VectorService {
         uri: Uri,
         compression: VectorCompression,
     ) -> Self {
-        let (protocol, endpoint) = uri::protocol_endpoint(uri.clone());
         let mut proto_client = proto_vector::Client::new(HyperSvc {
             uri,
             client: hyper_client,
@@ -83,14 +34,27 @@ impl VectorService {
 
         Self {
             client: proto_client,
-            protocol,
-            endpoint,
         }
     }
 }
 
-impl Service<VectorRequest> for VectorService {
-    type Response = VectorResponse;
+#[derive(Debug)]
+pub struct PullEventsRequestError {
+    pub status: tonic::Status
+}
+
+impl std::fmt::Display for PullEventsRequestError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str("Failed pull_events grpc request: ")?;
+        f.write_str(&format!("{}", self.status))?;
+        Ok(())
+    }
+}
+
+impl std::error::Error for PullEventsRequestError {}
+
+impl tower::Service<proto_vector::PullEventsRequest> for VectorService {
+    type Response = tonic::Streaming<proto_vector::PullEventsResponse>;
     type Error = Error;
     type Future = BoxFuture<'static, Result<Self::Response, Self::Error>>;
 
@@ -105,24 +69,19 @@ impl Service<VectorRequest> for VectorService {
     }
 
     // Emission of internal events for errors and dropped events is handled upstream by the caller.
-    fn call(&mut self, list: VectorRequest) -> Self::Future {
+    fn call(&mut self, list: proto_vector::PullEventsRequest) -> Self::Future {
         let mut service = self.clone();
-        let byte_size = list.request.encoded_len();
 
         let future = async move {
             service
                 .client
-                .pull_events(list.request.into_request())
+                .pull_events(list)
                 .map_ok(|response| {
-                    emit!(EndpointBytesSent {
-                        byte_size,
-                        protocol: &service.protocol,
-                        endpoint: &service.endpoint,
-                    });
-
                     response.into_inner()
                 })
-                .map_err(|source| VectorSinkError::Request { source }.into())
+                .map_err(|status| -> Self::Error {
+                    Box::new(PullEventsRequestError{ status })
+                })
                 .await
         };
 
@@ -136,7 +95,7 @@ pub struct HyperSvc {
     client: hyper::Client<ProxyConnector<HttpsConnector<HttpConnector>>, BoxBody>,
 }
 
-impl Service<hyper::Request<BoxBody>> for HyperSvc {
+impl tower::Service<hyper::Request<BoxBody>> for HyperSvc {
     type Response = hyper::Response<hyper::Body>;
     type Error = hyper::Error;
     type Future = BoxFuture<'static, Result<Self::Response, Self::Error>>;
